Rating view started

+++ Rating/Rating.Common/DataModel/Player/DriversRating.cs
@@ -5,10 +5,10 @@
     public class DriversRating
     {
         [XmlAttribute]
-        public double Rating { get; set; }
+        public int Rating { get; set; }
 
         [XmlAttribute]
-        public double Deviation { get; set; }
+        public int Deviation { get; set; }
 
         [XmlAttribute]
         public double Volatility { get; set; }
